Inserted links should work as internal wiki links
With document.execCommand('CreateLink',..) it will be simple external
links.

Parsoid converts the links to internal link when there is an attribtue
rel="mw:WikiLink"

To add that attribute, we need to create the whole link html and
insert/replace at curser/selection.

We can not use InsertHTML command since it is not supported in IE

<a href='Link' rel="mw:WikiLink">LinkText</a>
becomes
[[LinkText|Link]]

Without the rel attribute parsoid gives [Link LinkText]

Change-Id: Iab240509925fd278111fff492b414a65314afb8f