Insert links to source article and guideline in JS
This changes the links from being inserted using message
parsing to explicit creating of <a> elements in jQuery.

This also makes sure that the source article link
points to what is configured as the source wiki
and the guidelines link points to the target wiki.

Followup to I4565f87afe614bdc7799096fb064caea0903f816

Change-Id: I84c55d3ddbe4c341c364201e3ff6f53ad2cfdf36