Rewrite the progress calculations to provide more data
This way the end users can calculate whatever ratio they want
based on the mt, human and any weights, which has been normalised
to be in range [0,1] for convenience.

The object can be extended to include new items if needed.

Updated the progress bar tooltip to say the relative mt percentage
(ignoring untranslated sections).

Change-Id: I5660c74a9a3c6aeb3744a85f707531bbbdc02217