Post MT: Click has same effect for section pairs
After MT, clicking on source section has same effect of clicking
on target section. That means, clicking on the source section opens
up the MT tool card just like a translated target section.

Indirectly this fix the following issue:
When a link is highlighted, it is not possible to remove the highlighting
by clicking to the source paragraph regular text.

Change-Id: I34b8b1d9fb768a62fbdcaf7694bd4a95deef1b8a