While clicking on redlinks in source article, make the link red
The source article shows blue links for all kind of links. But it is when
user click on a source link, we get to know whether it is a redlink.
Mark it as redlink by adding a class redlink.

Ideally, we need to identify red links when page is loaded, but that is
bit complicated and involves API hit etc.

This solution seems to be 'better than nothing' approach.

Source link card does not appear when clicking on such links.

Change-Id: I7aa777e2bbabad76e068d0991bb185d5615786e2