Move all tools inside single container
Renamed cx-tools--container to cx-tools because we should not have
BEM classes which do not have corresponding base class. Also renamed
some other classes to better match BEM conventions.

Used cx-card--fixed class to mark cards which are always visible.

Instead of storing a reference to a wrapper div and using .find( input )
on it, store a reference to the input element directly.

Change-Id: I8a245226c416ced06720edbd6b9892bda64dac8b