Make all of Content Translation a beta feature
The existing beta feature for entry point has been absorbed into
a new beta feature that also covers the visibility of the special
pages. Using a combination of SpecialPage::isListed and check in
Special::execute to have the special page act as if they do not
exist for users for whom the feature is not enabled.

Added a global for the beta feature use, so that developers can
turn it off for example. As before, if BetaFeatures extensions
does not exist, acts as if the global was false.

Removed old beta feature images and added new ones.

Bug: T76175
Change-Id: Iff4ff583fe434ef7f15bc7af53e093ce1490a95c