Make sure translator see license text even if use manually created URL
When users try to access the translation view directly for
an article that they have not explicitly started before,
instead of being directed to the translation view,
they will be directed to the translation dashboard.

There, the New translation dialog will open and they will be able
to start the translation with a single click
(after reading the legal terms, as we all usually do).

If draft URL parameter is present, this step is avoided because drafts
are created after following the above step. In case translators try
a URL with an invalid draft, there will be draft validation (it is
a separate task).

A domain level cookie with short expiration time is used as a token to
indicate that the translator went through the normal translation
workflow. Note that this cookie is not host level, but at domain level,
domain being .wikipedia.org for Wikipedia. This allows CX to redirect
users between source and target language wikis.

Make sure to test this in a multi-language wiki setup and with
the following line in LocalSettings.php:
$wgContentTranslationTranslateInTarget = true;

Bug: T76178
Change-Id: I640ebdd801f99f85a27733b8b30dede21f0a4e8c