Override beta feature preference if Special:CX accessed with valid token
Translations in CX involve two languages. When creating a new translation,
CX redirects the user to the target Wikipedia so that articles are created locally.

This means that a user in the Spanish Wikipedia CX dashboard can start a translation
from Spanish to Catalan, and as a result he will be redirected to Catalan
Wikipedia's CX translate editor. This redirect allows CX to present an integrated
view of all translations (the user does not have to figure out to which site to
start a translation from), but avoid cross-wiki saving issues.

When CX is isolated as a beta feature, the redirect may be problematic since
the user needs to have the beta feature enabled for each of the languages
involved. If that is not the case, the user will get a page not found error
instead which will break the user experience badly.

To avoid this issue, we allow access CX even if the beta feature is disabled only
in the case of being part of the described redirection. The rationale is that
the user has already enabled the beta feature on a site,
and is clearly expressing the intent to start a new translation
for which reaching CX will provide less surprise than getting an error message.

Bug: T78001
Change-Id: Ia54f8912ef5f9dad20af5d2ebaaa966df767de67