Do not create links for the selected text in a link
When partial link text is selected, it is interpreted wrongly as
selecting translation text and clicking on source link. Resulting the
selection converted to a link. That makes a link inside link.

The problem was that there was no explicit check whether you are clicking
on a link in the source column or the target column.
This commit adds that check to solve the problem.

Bug: T85930
Change-Id: I2ea0480ceb3742e48426b0d01814c00e2cf16ad4