Links: Selecting partial or full link text does not make a valid selection
Check if the text selected is text of a link. If so, that substring
of link text is not a valid text of any link card related actions.

If we treat that text as valid selection, the link module allows creating
a link around that text - resulting link inside another link.

Similarly, if that text is valid, the "Add link" button is appearing
in target link card.

Bug: T86511
Change-Id: Iac02a1cf3d4eaa969e009dd6b5a4790802366823