Fix incorrect redirects to new translation dialog
When accessing /wiki/Special:CX, the cookie would get set with /wiki
path, but we are redirecting to /w/index... so it doesn't match.

Renamed prefix to path. $.cookie does not have a prefix option. I
believe someone was reading documentation for mw.cookie while doing
this. Used the global variable that already exists.

Change-Id: Ifbede3eb58469503abef2b9af89b9e443408fed2