Expose published translation with source-target URL pairs
Introducing an API to expose the published translations information
so that people interested in analyzing the translation pairs
(can't say parallel corpora in strict sense) get minimal information
to work with.

Sample output

"result": {
    "translations": [
        {
            "sourceTitle": "SourceTitle",
            "targetTitle": "TranslatedTitle",
            "sourceLanguage": "es",
            "targetLanguage": "pt",
            "sourceURL": "//es.wikipedia.org/wiki/SourceTitle",
            "targetURL": "//pt.wikiedia.org/wiki/TranslatedTitle",
            "stats": {
                "any": 1,
                "human": 0.45176110260337,
                "mt": 0.54823889739663,
                "mtSectionsCount": 2
            }
        },
	....
    ]
}

This is not a final version of the information we wanted to expose,
but just starting work on it.

As we improve our infrastructure we may able to provide
more useful information.

Bug: T87353
Change-Id: I0094102b7f04bd249e7f69502737f671755165c1