Change the status to "deleted" while deleting instead of removing record
Avoid losing all history of a translation when it is deleted.
The draft content if any is still hard-deleted, as is the record in
translator table.

We do not allow translating once a translation is published. But there
are cases where a translation is published, and translator does some more
translation that is not published. This translation will be in draft
state and can be removed, causing data loss about translation.
This data is important for providing translation pairs, proper stats etc.

This need a table change as follows

ALTER TABLE cx_translations
MODIFY COLUMN translation_status
ENUM('draft','published', 'deleted');

Bug: T89442
Change-Id: Id100294ca2c7993708bb7ca94350b2e93297677b