Database: Set targetURL only when status is published
Currently we update translation record in CX database for save and
publish actions. For both, we update the translation URL.
But translation URL update should happen only when publishing happens.
Otherwise, we just calculate a URL and add to record, a URL that
does not exist.

Changed the upsert query to save the translation to separate
insert and update queries.

Bug: T91365
Change-Id: I4cd9aa09bcac5318e3ee93e32abe1bf2d49be331