Show the newarticle campaign in the VisualEditor
VisualEditor is loaded in the frontend, so the campaign module
needs to be loaded using a frontend hook.

This commit adds a simple module, which is loaded if the page is new,
the campaign is enabled and the user didn't enable CX, and which loads
the actual campaign module at the ve.activationComplete hook.

Bug: T95269
Change-Id: I9198ea6edd45d0264650145abb62087cceb9bf87