Publish preprocessing: Remove the empty br tags inserted by Firefox
Testplan:

Translate some article. Press enter in between a paragraph text to
leave line breaks. In a header, remove the template text, and type something new.
Publish.
Expected result: The line breaks in paragraph should appear in published text.
And there should not be any <br> tags in the header.

Bug: T99851
Change-Id: Ia91f2f2fbe42a2a329df0874357147653859fe7c