Fixes when language code and domain differ
The toolserver tool handles redirects itself, so it was not
changed in this patch. All instances of lllang parameter were
changed.

Testplan:
Translate Tapethok from en to bho

Without this patch, the user is not notified that the page
already exists in the language.

Translate LaTeX from simple to bho
Without this patch, user will see all links in translation
template as missing. With this patch, links will be adapted
correctly.

Bug: T99888
Change-Id: I2d5021c27f8b03563ef5a16723bed3ea83d0e701