Convert unadapted links to plain text while publishing
Ideally, we will require a validator and cleaner from each tool
module before publishing. It is a bit architecture work we have to
do in near future. For now, this is done in a simple way.

Testplan:
Translate LaTeX from en to es. Click on first section, note the
unadapted links in translation. They will be in gray color. Publish.
The published article should not have them as links. They should be
plain text.

Bug: T78133
Change-Id: Ib530fd95c7934f422a6b5850cd06fd861a9987cb