Separate and isolate source and target title in the dashboard
Without proper isolation the titles can appear in the wrong order
in RTL wikis.

Even if both titles are in LTR, in RTL UI the source must be on the right
and the target must be on the left.

To test, create a translation with a different target title,
type something and save it, and then take a look at the dashoard.

Bug: T105903
Change-Id: Icadae886e6060f11e57cb95cdce593bca39fac43