Do not show the article selector unless all params passed in URL
It seems that showing this dialog even with partially filled params
(for example, contribution menu entry point fills the target lang param)
creates confusion to users and they re-type title and end up in creating
a fresh translation that overwrites an in-progress translation.

Target title is not validated for this, since it is optional.

Bug: T104702
Change-Id: Iff73f4bedbb1038d70ca26e087550ddb64261220