Improve reference adaptation to avoid parsoid errors
The current reference adaptation does not guarantee that the adapted
reference has a valid data-mw attribute. Parsoid demands a valid id
in data-mw.body pointing to an item in references section. Not having
that causes parsoid errors during HTML to wikitext conversion.
It does not cause parsoid to fail, but generates articles with errors
as shown in Bug T107252.

This patch improves and cleans up the parsoid adaptation process.
It also makes sure a valid data-mw is present in the adapted reference.

Testplan:
Translate Vaporwave article from English to Spanish. Use MT for all
sections and publish. The published page should not have any errors
about references. Compare it with
https://es.wikipedia.org/w/index.php?title=Vaporwave&oldid=84053572

Make sure the reference card is working for all references and
adding reference at cursor position also works.
Repeat the test for translating Attack & Release article from en to es
and compare result with
https://es.wikipedia.org/w/index.php?title=Attack_%26_Release&oldid=84052884

Bug: T107252
Change-Id: I0ab01b682a54ac3a96a52068740af85964e99807