Apply selected action on Translation card to the current paragraph too
Testplan:
Translate a section. Change the translation provider choice using
the MT tool card. The choice should be applied to the current section.

If choice is an MT engine, the current section will have the machine
translation.
If choice is 'Use source', the current section will get copy of source
section.
If the choice is no machine translation, the section will be cleared.

There is a bug about what happens if the section has user's manual translation.
This commit does not change any behavior on that. It just overwrites it as we do now.

Bug: T78143
Change-Id: Ief487458de9743a10d99e441bcf06761e6db20ff