Remove CX generated attributes from content to publish
Never leak them to articles.

Testplan:
Translate article: ISO_3166-1 from en to any language. Add the
table in the source article and publish. You will see that the
wikitext of the table as

{| data-cx-weight="11017" data-cx-state="source" data-source="mweg"
class="wikitable sortable" data-seqid="162" id="cxmweg" contenteditable="true"

Which is bad. Now with this patch you should see:

{| class="wikitable sortable" id="cxmweg"

Keeping the id for now, because it need more serious testing.
I think parsoid will require it for templates or reference handling.

Change-Id: I4d050330540567fbe0d5e0c36d0839fb14257d92