Enable scrubWikitext=1 in HTML to wikitext conversion using parsoid
This is a flag to fix up HTML in a way that results in cleaner wikitext,
not worrying about clean round trip.
CX does not care about round trip since it creates articles, so we can
try to generate optimal Wikitext as much as we can.

See https://www.mediawiki.org/wiki/Parsoid/Normalizations

Change-Id: I0cb7df2f4e3992c468673072d8204ec047586c3a