Convert the language domain code mapping to a global variable
Until now the same data was written in two places - the PHP SiteMapper
and the JS SiteMapper.

This patch moves the data to the global configuration variable
$wgContentTranslationDomainCodeMapping,
which makes it written only once, without having to sync
two files every time a change is needed.
It also makes it customizable per site.

Change-Id: I9cda0ef692a8bc29caa02e70fe42e50cbf5fee44