Stats: Separate total and language graph, and show draft count in graph
This is in preparation for showing deletion stats and weekly trend
graphs.

Showing both total and language specific stats is difficult to
understand because the difference is widening fast. So separated them.

ApiQueryContentTranslationLanguageTrend now return cumulative count
for translations in progress as well. Soon it should retun
cumulative deleted stats too.

Bug: T105192
Bug: T90538
Change-Id: Ifd4bd0e3a92ac022428edf3915552ea99142dbfb