Discard draft: Validate against last updated translator
Validating against'startedby' will prevent the translator deleting
their translation in case same title was tried to translate by another
translator in the past.

Of course the last translator can use that title only after the previous
translator either discard or publish the title.

This change makes the condition to show a translation in somebodys dashboard
and condition to delete the same.

Bug: T97394
Change-Id: I53da74bf9d254f7c7533441f7b45524dfa3757c9