Make interlanguage link entry point gray in Monobook and Modern
Vector needs different CSS selectors for the interlanguage links list,
and overrides their default style. Till now this was worked around
using a super-specific CSS selector to set the links' color,
but this selector doesn't work in other skins.

This patch makes a simple default selector for other skins,
and puts the super-specific selector in a Vector skinStyle.

Tested with Vector, Monobook and Modern skins.

Bug: T113515
Change-Id: I1bb057d48f0e61597abd3ef373726325c5210e7a