Suggestions: Public task lists based on category and their paging
A script to create a suggestion list based on a category is added.

Pages that exist in the source language and not in the target language
will be added to that list. There is no UI to create or manage
this list. For now it's just a maintenance script.

Category based task lists are presented as collapsed list in the UI
with just two items. The expanded list can be collapsed, too.

Since we are getting more lists, infinite scroll behavior has changed
significantly according to T115008.

When you scroll, if there is any expanded list present, more items from
it will be loaded. If there is no expanded list, when you scroll, just
the scroll happens.

There is no infinite scroll for the default suggestions
(the featured list suggestions, also called "fallback suggestions").
Instead there is a trigger to refresh that list with new items.

Since lists need to manage their own offset and continue params,
they were removed from class level and added as member of list objects.

Bug: T115008
Bug: T115006
Change-Id: I3eb29a2337f2ddbb2bc9bcf86f8efcfa532b0984