Fixed a bug that prevented suggestions from showing
Personalised and public suggestions are merged together, but only
if you had personalized suggestions, i.e. you at least once
favorited a suggestion. If you hadn't, the complete list of
suggestions was not populated at all. This seems an oversight in
the code that is easy to fix, as both personalized and public
suggestions are guaranteed to be arrays (even if empty) which can be
passed to array_merge unconditionally.

Obviously, I had tested favoriting and discarding when it was
implemented and never noticed this problem in my testing.

Bug: T118194
Change-Id: Ibcd8a336f6c2b9da50c05f3c694738ced16471dd