Do not rely on draft param to restore a draft
In case of reloads at translation view, there is no draft Id in the URL
and it will result presenting a fresh translation view. Even though,
CX had saved all translations in database. This can be interpreted
by users as 'data loss'. The page refresh might be accidental too.

With this patch, we dont rely on draft id in URL to restore the existing
translation. But we will query CX to see if there is a translation for
given source title and language pair. If so fetch the translation content
and restore.

References to draft id is removed, instead, the translation id found while
restoring is set to mw.cx.translationId. We need this for section level saving
of translation very soon.

Also cleaned up the Deferred usage to make this module more testable.

Testcase:
Start a fresh translation in CX. Translate a few paragraphs. Make sure
the content is saved by autosave or just press Control+S. Then reload
the whole page. The previously saved translations should get restored.

Bug: T120635
Change-Id: I157b0204425cfe18805ae0ecf86309a9df866e13