Do not pass draft id while resuming, remove references to draft id
SpecialContentTransation class now does not check the draft in
the request params to check whether user had read license agreement.
cxtoken or existance of the translation is used for this validation.

Refactored the click handler of translation list items so that it
will set the cxtoken.

Bug: T121116
Follow-up to I157b0204425cfe18805ae0ecf86309a9df866e13
Change-Id: I254e98deafc4a05a7bd0a576c21a0d4c7d4752d5