Remove hyphens from all languages when creating a gray link
The current code was removing hyphens only from accept-language,
such as en-us. Such hyphens can also appear in language codes
that come from other sources, so now they are removed from the whole list.

The special check for be-tarask is removed, because now any code
that begins with be-* will change to be and then split to
be and be-tarask.

Bug: T113945
Change-Id: I77c5a0380a8f20d5c1f0a898dd780e5bfbf75229