Do not fail updates on duplicate units
The schema allows duplicate units that only differ by timestamp.
In practice, for now at least, we are aiming to keep only latest.

This leads to a problem when there are more than one created by
race conditions. Fix the update method so that it does not match
multiple rows (updating them would violate the unique index) but
instead only one row. This prevents an error and saving succeeds.

Alter the find method so that it always picks the latest unit if
there are multiple, so that we return the latest saved translation
and that we also update the latest saved translation.

Preventing the duplicates in the first place is left as an follow-up.

Change-Id: Ib700be6b773dfb2035417159a88fd9bce0df4739