Show the save progress indication
In the old approach we were using mw.cx.translation.save event for this.
But that is not reliable since it is not guaranteed a real API call happen
for that event. We check if there is something to save and then save if any.
So if nothing saved, there won't be mw.cx.translation.saved event causing
the 'Saving...' indicator never disappear. So using mw.cx.translation.save
for the purpose of progress indicator at UI is problematic.

So in this commit, we introduce a new event specifically to indicate
a real save API call is starting.

Bug: T128707
Change-Id: Ia6346c37ccee7b2a65d09473b5e4007ee21e7a74