Better announce new optional MT services available
Bug T128059 has an additional requirement that the green indicators
are never shown again once translator has chosen the new provider.
This is not done in this commit. It requires a way to persist
the provider preference. To be done in a different commit.

Note that the expected configuration to set a non default MT in cxserver
configuration with this patch is:

mt:
 defaults:
   'en-es': source-mt

Bug: T128059
Change-Id: I4bf8635a03ce21e334adbfdb9fffbb9963179699