Do not count a section restored as orphan as orphan again
If a section was not able to restore and if there is a successfully
restored last section exists, we add it after the last restored section.

But once that is done, we should not still add that section to orphans
array. There was a logic error in the code. This commit fixes it.

Change-Id: If62429a1494d746e37c9fadf86c93ac01a21fba9