Templates: Use wikidata to find template in target wiki
The existing check was simple title search in target wiki. This
was problematic and not helpful. Replace it with langlinks api
which will resolve the titles using wikidata connection information.

Testplan:
Translate Giovan Battista Perasso from it to en. Inspect the reference
"Gattera 2007, pagg..". The translation should be using 'Harward
citation no brackets'

Bug: T126957
Bug: T125579
Bug: T112739
Change-Id: I346d5653b7f6a4756f59e99881242d1cf2d17fb6