Remove click handler for target language selector in Special:CX
We were doing a lot of job to remove the target language selector
before replacing it with a new one. Unfortunately we forgot to
remove the click handlers that were creating all the old ones
each time the target language selector was opened.

This created two bugs.
1. focusing the search field was not working when more than one
target selector existed.
2. selecting the previous source language as target language was not
possible because we were showing the old target language selector
where it was not listed.

Bug: T137184
Change-Id: I45539024b0e3473851894809863413118dcdc260