Avoid template name only adaptation for Cite web
Cite web template usually has mandatory parameters like url and title
We can do language pair parameter mapping, but not in all cases.
For rest of the languages, when template name alone gets adapted and
parameters do not, adaptation produces wrong wikitext. It is better to
use 'Cite web' as the template name in such cases since that template
might exist as such in those wikis.

In this commit we instruct the template adaptation to do that using
the common configuration.

Testplan:
* Translate Eleanor Parker from en to no without this patch.
* Translate the section "Source: "Eleanor Parker". IMDb.....".
* Inspect the translation. You will see the template name changed
  to Kilde www, but parameters unchanged. With this patch you will
  see the template name changed to Cite web.
* Publish the article and make sure the published article is clean.

Change-Id: Idc58cc06951e080943fcc8ad1ba1e0210938ef2a