Source filter: Do not remove references list
In pt.wikipedia.org, it was noticed that the references list is
created using {{Referências}} template. It creates an <h2> for
header and then a <div> containing reference list.

While filtering out block level templates, the <h2> is removed since
it is not a references list and while doing that the elements connected
with same about attributes are also removed, resulting removal of
reference list inside the <div>.

While doing template removal, we check if the template is reference
list or its children are reference list. If so, we do not remove
them since reference list is must for successful reference adaptation
and publishing. But we missed to check if one of the related element
with same about attribute is reference list.

This patch corrects it and keeps the reference list.

Bug: T142983
Bug: T142753
Change-Id: I89e24763f1df0fb93bb247d40032307c1016ca8d