Translationview rewrite based on OOJS, OOJS-UI
The three column layout and the initialization code for
the components were rewritten using OOJS-UI. Rest of the modules
left untouched. So no functionality change expected.

The UI code is now organized under the ui folder with mw prefixed
files and modules, dropping 'ext' prefix so that the classes are
easily greppable with its names.

Another notable change is the infobar which used to be inside the header
module is now separated.

This is a beginning of changes that leads to further modeling of our
classes for VE integration.

Bug: T149163
Change-Id: I07cde72f9b4a7daa45415b4af16f086f6c172046