Add configuration option for cookie domain
Our current implementation defaults to a domain which is the
current domain stripped of first subdomain, if any. While this
works for testing and for Wikipedia, it does not work Wikimedia
Labs because it is listed in [1] as public suffix. That means
labs projects cannot set cookies to wmflabs.org, hence the
setting of the cookie failed. Since this behavior is hard-coded,
I added a new configuration option to explicitly set the wanted
domain. For now it defaults to null to keep the custom behavior.
For Wikipedia we should set this value explicitly to wikipedia.org
so that we can require setting this value explicitly.

I also changed the code to use mediawiki.cookie instead of $.cookie
to make it simpler. There was no explicit dependency to
jquery.cookie, but I now added it for mediawiki.cookie.

I tested this manually on my development wiki. To configure, set
$wgContentTranslationSiteTemplates['cookieDomain'] = 'something';

[1] https://publicsuffix.org/list/public_suffix_list.dat

Bug: T149879
Change-Id: Ife52adde319c6cae52cf523cb18e8d0103dae519