Templates: Editing support and adaptation controls
1) Templates: Do not remove them from source

We still cannot keep everything in source article. Navigation
boxes is one example. The configuration to black list certain
elements from source is given in a json file.

2) Template control card for block level templates

The template card provides 3 options for block level templates
1. Adapt the template to target language(default). If adaptation
fails, mark the template as unadaptable, section not included for
publishing
2. Keep original template from source language.
3. Skip. Just ignore the template.

The TemplateTool class works only with the template definition
and avoid directly dealing with the DOM element.

There is no editing support for any of the above options in this
iteration. Because of that, MT card, formatting, link cards wont be
shown for this kind of sections.

Updated the tests so that it uses template definition and avoid
any manipulation on DOM. Makes it easier to understand and maintain.

Split the Template card from ext.cx.tools.template module

3) Block and inline template editor

* Created mw.cx.Template class to abstract a template
* Created an editor with template params and values
* Added an API to convert fragments of html and wikitext
  in both directions
* Once template edited, rerender it in the translation column
* 'Use source' and 'skip' won't have template editor.
* If target template does not exist, no editor provided.
* If template data is not found for target template, no editor
  is provided.
* Misc bug fixes in other modules required for the editor
* Machine translation will be in another iteration
* References list usually has high number of templates. To avoid
  processing all of them when the references list section has added,
  only the references present in translation is processed.

Bug: T143121
Bug: T144336
Bug: T144004
Change-Id: Ie87725b3a2a26894093e1d51b9fdafbc393fba2c