Fix 'Main Page' appearing while auto-adapting templates
"Main Page" is frequently shown when auto-adapting templates that use
{{PAGENAME}}, instead of the actual page name. This is because the
templates uses the name of the current page and that information is not
supplied to parsoid by u while doing wikitext to HTML transformation
from the CX Template editor.

Pass target title to parsoid to fix this.

Testcase:
Translate 'Kevin de León' from en to es. In es, the infobox uses
template 'Ficha de autoridad'. It is a wikidata based template and needs
a page name. Without this patch, you will see Main Page coming as the
adapted template. With this, the whole infobox should get rendered.

Bug: T154297
Change-Id: I0f2b9b633c214aec7ee2d45e3c25e19cb7ccc527