OOJS based refactoring
New data models out of the existing code. These classes are based
on OOJS and part of VE migration code refactoring.

Concepts:
* Any translatable element which has correspondence
  in target language, automatically adaptable or manually translatable
  and editable are defined as TranslationUnit.
* The translation context is available in Translation Units and other
  classes by dependency injection.
* Data manipulation, adaptation are done in DataModel classes and made
  independent of UI or contentedtiable elements. Changes in contenteditables
  are updated in datamodels whenever content changes.
* A translation unit can have zero or more translation tools which are tools
  shown in translation column.
* All translation units are wrapped in <section> tags

This is just an iteration and not a replaceable code for the CX in production.
Once we have done with this reorganization, we can remove old code. Till then
old code is not removed. The new code is easily identifiable since the file
names are prefixed with mw.cx. Old code is prefixed as ext.cx or present
in folder named 'legacy'

Use version=2 param in URL to use this version of code.
Also introduces wgContentTranslationVersion configuration setting for same purpose

Bug: T152586
Change-Id: I4b4058fed953061e81dabaf5f08ab3e933ca01b9