Make sure the click event is always emitted to all translation units
Section and Template translation units were not getting it since
SectionTranslationUnit overrode the onClick method.
Now that Section get the event, we don't need to register tools for
SentenceTranslationUnit.

All these event emitting pattern need to be better written with
documentation and assert that they dont get lost to consumers

Change-Id: I12b88ee564276e54f7d481c86b44d051f0ddf546