Big rework to cx_translations table
Original idea was that one article (name, source language, target
language) can only be translated once. Hence we have unique index
on that triple. However, articles get published or drafts deleted.
In these cases we use the columns started_by and updated_by. However
this was updated by code that was fragile, not well tested, and it
was losing data. Now we have about 500 rows where we cannot be sure
which one of started_by or updated_by would indicate the current
owner of the article, if either. The translators table is supposed
to keep track which user has which drafts, but there are many
instances where two different users have access to the same draft.

To avoid all these problems in the future, change the unique index
to include translator_id and enforce the single-active draft
restriction in the ui and saving. I updated a lot of code to use
this 4-tuple piece of information. In addition I created a new class
TranslationWork to encapsulate the original triplet to make the
code nicer to read.

This will temporarily block all access to existing drafts where
started_by != updated_by until those are sorted out. It will also
prevent any more ownership changing of existing rows until the
patch to update the index is applied. After that this code will
automatically start creating new rows for different users, if
they don't conflict with on-going drafts. The schema change
should be applied after this code is deployed. And later on
the temporary support for old index can be removed.

I also added an exception to Translator if global id cannot be
found. We seem to have a few dozens of zeros in our tables, some
related to logged out users (even though we use assertuser in the
API calls).

In future updates, the updated_by column should be dropped. Also
the cx_translators table could be replaced with an index in the
cx_translations table, as they right now duplicate information.

Bug: T146450
Change-Id: I4ecf067ed75a4b0b7f26e6f00c5d6e23ae54e503