Add mw.cx.TargetArticle class
MediaWiki specific validations, publishing, success and error
handlers. Supposed to replace ext.cx.publish module in old CX

The success/error message system is not ready so those can't be
tested now - to be done in follow up.

Captcha handler and title conflict errors need to present a dialog
to user to resolve it - to be done in followup.

Change-Id: I5eab6fec79895591ea0b404196fb30f17452f75a