Fix PHP error screen while landing in a wiki as anonymous
After the source language and titles are selected, the user will be
redirected to target wiki. We cannot assume that the user is logged
in there. The isExistingTranslation method will be triggered before
we check if the user is logged in or not. That will trigger a PHP
error in Translator class saying 'User account is not global'.

This patch present that error nicely.

Change-Id: If8785b95695a8967727dbdc30a75e0a85ac0afdf