ApiContentTranslationConfiguration: fix configuration merging
Common configuration was not included if language pair specific
configuration exists (e.g. ?action=cxconfiguration&from=fi&to=en).

This is now fixed and the code is also simplified. In CX1 this was
not a visible issue, but in CX2 the code was failing because
`removableSections` key was not present.

Change-Id: Ie3be911ad5c378ca84738b8538111bf497633bab