Fix the container calculation for template editor
For inline templates, it was possible to have a container
for editor evaluated as a child of a section. Since anything
under section is saved, the whole editor code may get saved.

Also the previous code might give more than one match for the container
resulting multiple editors.

This fix corrects this behavior.

Testplan:
Translate Варна from ru to ba. Click on the section under Климат, it has
two big tables, both are templates. Double click to open the editor.

Without this template, the editor's position will be wrong and sometimes
the target column of editor won't even appear. Also the whole section
alignment will go wrong.

With this patch the editors should appear aligned and in correct position.

Note that there is a delay in getting this template editor rendered since
it has lot of fields.

Bug: T163105
Change-Id: I4094c81c08a0f6674664787ae90d2eb446c36c6e