CX2: Give correct page missing and other info from LinkCache
page.missing property won't be a boolean property from API response.
If the page is missing, the missing key will be present in the response.
Made this same as VE version of LinkCache.
While at that, tried to capture all other information that this API
is fetching. Not everything is used, but might be helpful.

Change-Id: Ic3a5190f5754a8d3a218cf4611364f616a61da89