CX2: Add new link to target section from selection
When a text is selected from the section, a link card
is presented showing whether the selection can be changed
to link. If the selected text is a title in target link,
pressing add link button will change the selection to a link.

This link can be inspected using link card - by clicking on it.
It works just like any other links, except these new links won't
have a corresponding source link.

For the selections, which does not have a target link, card will
give option to change it to a missing link. This is not implemented
in this patch.

Bug: T162126
Change-Id: I52eb821ea6cab45f6152783adeb5d8149c2f29f2