CX2: Do not silently fail in case of initialization errors
If there is programming/js errors in any of the modules, CX
was not reporting it in console. The UI just shows loading
indicator. This changes fixes it. The error trace is given in
console and critical error is displayed on UI.

The initialization related fetches are now parallel.

To reproduce the issue, introduce a programming error, for
example, in mw.cx.dm.TranslationUnit constructor, replace
'onChange' with 'onChangexx' and reload the translation.
Without this patch, the UI just stuck in loading stage. With
this patch, a big error should be in UI and exact error in
developer console.

Added more logging.

Bug: T166770
Change-Id: Ibf241fdb9b8c9933b76324ed40143d336be0ce50