Templates: Filter out autogenerated nodes before copying data-mw
After a template is adapted, we update the HTML in translation view
and use the same for publishing.

For some templates, the generated HTML will be multiple fragments.
For example, a template like:
{{שדה תעופה|שם=Pokhara Airport|IATA=PKR|ICAO=VNPK|מפעיל=Civil Aviation Authority of Nepal (CAAN)}}
will produce HTML output as
<span data-mw..> // Span with template definition
<link rel="mw:PageProp/Category"> // hidden category
<link rel="mw:PageProp/Category"> // hidden category
<link rel="mw:PageProp/Category">
<table..> //The actual infobox table

The HTMl we need to pass to parsoid during publishing should be
<table data-mw>.. // data-mw copied from span
And all other elements should be removed.

While processing, we remove all these <link>. Also the copying of data-mw
is to the 'second' fragment. If we do the filtering after this copying,
data-mw will end up in a <link> node that is removed.

So in this commit, we do filtering first and copying later.

Bug: T169030
Change-Id: Ic5d9deca22ec2ad66a498f4dab565dcbe33a1fc3