Change action icons for list items in dashboard
Icons for actions are placed at the top-right corner of the element
An outlined star is used for items that are not in the "For later" list,
while a filled star is used for those added.
Color of filled star is changed from green to blue, to match design guidelines.

Bug: T160356
Change-Id: Ib544d610f272839bb4631dd536b80d4606578494