CX2: Force <base> to current wiki in target document
With I8367f9f2764, cxserver now returns a full document with the <base>
tag pointing to the source wiki, so the source document automatically
picks up the correct base URL. However, the target document's base URL
needs to be the URL of the current wiki.

Fix it up by removing all <base> tags from the target document,
then running fixBase() again. This ensures that the links in
each document (both in the surfaces and in inspectors) point
to the correct domain.

Change-Id: I886d022f67e0198fbd4877e17cbdae77679fcdd4